+++ app/src/main/java/io/itch/fr/quizgame/MainActivity.kt
@@ -3,28 +3,24 @@ package io.itch.fr.quizgame
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.viewModels
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
 import dagger.hilt.android.AndroidEntryPoint
 import io.itch.fr.quizgame.feature_start.presentation.StartPage
-import io.itch.fr.quizgame.feature_start.presentation.StartViewModel
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            TriviaQuizApp()
+            MyApp()
         }
     }
 
+    @Preview
     @Composable
-    private fun TriviaQuizApp() {
-        val startViewModel: StartViewModel by viewModels()
-
-        StartPage(
-            onStartQuizClicked = { /* Handle start quiz button click */ },
-            onViewHistoryClicked = { /* Handle view history button click */ }
-        )
+    fun MyApp() {
+        // You can use a Navigation component or a simple composable function for navigation
+        StartPage()
     }
 }