+++ app/src/main/java/io/itch/fr/quizgame/feature_start/presentation/StartViewModel.kt
@@ -1,9 +1,0 @@
-package io.itch.fr.quizgame.feature_start.presentation
-
-import androidx.hilt.lifecycle.ViewModelInject
-import dagger.hilt.android.lifecycle.HiltViewModel
-
-@HiltViewModel
-class StartViewModel @ViewModelInject constructor() {
-    // Add any necessary logic or data for the Start page here
-}